Get by filters

+++ src/main/java/com/alkemy/disneyWorld/service/implementation/CharacterServiceImpl.java
@@ -1,15 +1,15 @@
 package com.alkemy.disneyWorld.service.implementation;
 
-import com.alkemy.disneyWorld.dto.CharacterDTO;
 import com.alkemy.disneyWorld.dto.CharacterBasicDTO;
+import com.alkemy.disneyWorld.dto.CharacterDTO;
 import com.alkemy.disneyWorld.dto.CharacterFilterDTO;
 import com.alkemy.disneyWorld.entity.CharacterEntity;
-import com.alkemy.disneyWorld.entity.MovieEntity;
+import com.alkemy.disneyWorld.exception.ParamNotFound;
 import com.alkemy.disneyWorld.mapper.CharacterMapper;
 import com.alkemy.disneyWorld.mapper.MovieMapper;
 import com.alkemy.disneyWorld.repository.CharacterRepository;
+import com.alkemy.disneyWorld.repository.specifications.CharacterSpecifiaction;
 import com.alkemy.disneyWorld.service.CharacterService;
-import com.alkemy.disneyWorld.service.MovieService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,39 +20,38 @@ import java.util.Set;
 @Service
 public class CharacterServiceImpl implements CharacterService {
 
-    @Autowired
     private CharacterMapper characterMapper;
-
-    @Autowired
     private CharacterRepository characterRepository;
-
-    @Autowired
     private MovieMapper movieMapper;
+    private CharacterSpecifiaction characterSpecifiaction;
 
     @Autowired
-    private MovieService movieService;
+    public CharacterServiceImpl(CharacterMapper characterMapper, CharacterRepository characterRepository, MovieMapper movieMapper, CharacterSpecifiaction characterSpecifiaction) {
+        this.characterMapper = characterMapper;
+        this.characterRepository = characterRepository;
+        this.movieMapper = movieMapper;
+        this.characterSpecifiaction = characterSpecifiaction;
+    }
 
     @Override
-    public CharacterDTO save (CharacterDTO dto){
+    public CharacterDTO save(CharacterDTO dto) {
 
-        CharacterEntity savedEntity = characterMapper.characterDTO2Entity(dto);
+        CharacterEntity savedEntity = characterMapper.characterDTO2Entity(dto, true);
         CharacterEntity savedCharacter = characterRepository.save(savedEntity);                         //El metodo save devuelve la entidad guardada con id ya asigando
         CharacterDTO characterDTO = characterMapper.characterEntity2DTO(savedCharacter, false);   //y a esa entidad guardada la transformo en dto
 
         return characterDTO;
     }
 
-    @Override
-    public List<CharacterDTO> getAllCharacters() {
-
-        List<CharacterEntity> characterEntities = characterRepository.findAll();
-        List<CharacterDTO> characterDTOList = characterMapper.characterEntityList2DTOList(characterEntities, true);
-
-        return characterDTOList;
-    }
-
     public void delete(Long id) {
-        characterRepository.deleteById(id);
+        Optional<CharacterEntity> characterOptional = characterRepository.findById(id);
+
+        if (characterOptional.isPresent()) {
+            characterRepository.deleteById(id);
+        } else {
+            throw new ParamNotFound("Movie not found");
+        }
+
     }
 
 
@@ -62,6 +61,7 @@ public class CharacterServiceImpl implements CharacterService {
         List<CharacterBasicDTO> characterBasicDTOList = characterMapper.characterEntityList2DTOBasicList(characterEntityList);
         return characterBasicDTOList;
     }
+
 
     @Override
     public CharacterDTO getCharacterDetailById(Long id) {
@@ -74,10 +74,10 @@ public class CharacterServiceImpl implements CharacterService {
     public CharacterEntity getCharacterById(Long id) {
         Optional<CharacterEntity> characterOptional = characterRepository.findById(id);
 
-        if(characterOptional.isPresent()) {
+        if (characterOptional.isPresent()) {
             return characterOptional.get();
-        }else{
-            return null;
+        } else {
+            throw new ParamNotFound("Character Not Found");
         }
     }
 
@@ -91,7 +91,11 @@ public class CharacterServiceImpl implements CharacterService {
         characterEntity.setImage(characterDTO.getImage());
         characterEntity.setWeight(characterDTO.getWeight());
 
-        characterEntity.setMovies(movieMapper.movieDTOList2EntityList(characterDTO.getMovies()));
+        /*
+        if (characterDTO.getMovies() != null) {
+            characterEntity.setMovies(movieMapper.movieDTOList2EntityList(characterDTO.getMovies()));
+        }
+         */
 
         CharacterEntity updatedCharacter = characterRepository.save(characterEntity);
 
@@ -100,35 +104,14 @@ public class CharacterServiceImpl implements CharacterService {
         return updatedCharacterDTO;
     }
 
-    @Override
-    public void addMovie(Long id, Long movieID) {
-        CharacterEntity characterEntity = getCharacterById(id);
-
-        MovieEntity movie = movieService.getMovieById(movieID);
-
-        characterEntity.addMovie(movie);
-
-        characterRepository.save(characterEntity);
-    }
-
-    @Override
-    public void deleteMovie(Long id, Long movieID) {
-        CharacterEntity characterEntity = getCharacterById(id);
-
-        MovieEntity movie = movieService.getMovieById(movieID);
-
-        characterEntity.deleteMovie(movie);
-
-        characterRepository.save(characterEntity);
-
-    }
 
     @Override
     public List<CharacterDTO> getCharactersByFilters(String name, Integer age, Set<Long> moviesIdSet, String order) {
         CharacterFilterDTO characterFilterDTO = new CharacterFilterDTO(name, age, moviesIdSet, order);
-        List<CharacterEntity> characterEntityList = characterRepository
+        List<CharacterEntity> characterEntityList = characterRepository.findAll(characterSpecifiaction.getCharacterByFilters(characterFilterDTO));
+        List<CharacterDTO> characterDTOList = characterMapper.characterEntityList2DTOList(characterEntityList, true);
 
-        return null;
+        return characterDTOList;
     }
